backend/requirements.txt
amqp==5.1.0
aniso8601==9.0.1
async-timeout==4.0.2
Babel==2.9.1
billiard==3.6.4.0
blinker==1.4
celery==5.2.3
cffi==1.15.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
Deprecated==1.2.13
Flask==2.0.3
Flask-BabelEx==0.9.4
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
prompt-toolkit==3.0.28
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
redis==4.2.2
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.32
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.0
WTForms==3.0.1
